uint64 -> int

+++ Kernel/include/proc.h
@@ -60,7 +60,7 @@ typedef struct {
 
     //Informacion de los hijos:
     int childrenAmount;
-    uint64_t childProc[MAX_PCS];
+    int childProc[MAX_PCS];
     int blocksAmount;
 
     int isYielding; //1 true, 0 false
@@ -78,7 +78,7 @@ typedef struct {
     char my_prio[16];
 
     int childrenAmount;
-    uint64_t children[MAX_PCS];
+    int children[MAX_PCS];
     // Podrías incluir file descriptors si querés: los ids nada más.
     uint64_t fileDescriptors[MAX_FD];
     int fileDescriptorCount;  // Número de FDs válidos
